[INTERNAL] sap.ui.fl: clear prepared maps on component load
The information can be out of date, so as a workaround the prepared maps are completely cleared on component load.
The proper solution will be implemented with the new fl lib restructuring
Also the max layer filtering was moved to a later stage, when the first map gets prepared.

BCP: 2270000316
Change-Id: Idee0f1139d999f82b3e25d260f3a72688bc3b678