[FEATURE] mvc.XMLView: make "content" aggregation of XMLView bindable
The XMLView is extended by having its "content" aggregation to be bound.

When the "content" aggregation is bound, the controls that are in the
"content" aggregation are used in its renderer instead of the parsed
content from the XML nodes because the "content" aggregation is changed
dynamically by the binding updates.

This also requires the removal of the preservation of already rendered
DOM elements in order to update the DOM elements correctly after a
binding update. Without DOM preservation, no XHTML or SVG node can be
supported in the binding template of a bound "content" aggregation.
Using XHTML or SVG node in the binding template of a bound "content"
aggregation causes a runtime error.

JIRA: CPOUI5FRAMEWORK-349
Change-Id: Idb86b27d306e981b7fb60d1bc41e6eb46828af7d