[INTERNAL] Table: fix broken table layout in fixedLayout="Strict"
When there is only 1 visible column in the table and this column is larger than the table, sapMTableLastColumn style class is added to the 1 visible column to avoid content overflow.
When the table becomes larger and hidden columns are visible due to the resize, then the sapMTableLastColumn must be removed, since this style prevent the dummy column to have a width.
See the Table.less file for the style definiation.
We now remove the mentioned style class when there are other visible columns in the table.
BCP: 2270021215

Change-Id: I7fdc6c80e53e5bd0f552fb7e1acf51feb14e1c06
(cherry picked from commit c8c1bb63885de8e0fb428c3f51bb7463ab7b2a78)
CR-Id: 002075125900000492192022