[FIX] sap.ui.unified.Calendar: refactor focus handling
Problem description:
The switch button used for changing between date and time
views in the sap.m.DateTimePicker is unreachable via keyboard.
Additionally users could not reach the action confirmation
buttons in the footer of a sap.m.DatePicker popover via keyboard.

Solution:
The custom logic for forward and backward navigation
via tab is removed from our date and time controls.
Instead the control parts are added into the DOM in the
sequence of navigation. The display stays the same due the
to elements styling. All elements in the navigation sequence
are now focusable by default and the tab chain is left to
the browser to handle.

Additionally there is another fix for the sap.m.DateTimePicker
control. Previously when a user chooses a date from the calendar
the focus wasn't moved to the next time picker input field.

BCP: 2270014669
Change-Id: Ia8298b011bfbc27f9262ce27c3a58f337c990975