[FIX] sap.uxap.ObjectPageLayout: Fixed inconsistent rename information
Problem: When there is only one SubSection in an OPL's Secton, and it
has an empty string as a "title", the "title" of the Section is
displayed (despite the fact that the SubSection is promoted in this
case).
The UI Adaptation now has new feature - Visualization, which shows some
information about what has been changed, displaying the previous "title"
of the Section and the new one.
However, in this particular situation, as the "title" of the SubSection
is an empty string and the actual "title" of the Section is shown, in
the RenameObjectPageSection we provide the wrong control for renaming
(the SubSection, instead of the Section). That's why the Visualization
feature of the UI Adaptions, shows in this case that the "title" was
changed from empty string to the new one. This is confusing for the
user, as he sees the "title" of the Section and it is not an empty
string.

Solution: In case, when the "title" of the Section is shown, despite the
fact that the SubSection is the only one (but it has an empty "title"), in
the RenameObjectPageSection we provide the Section control for the rename
change.

BCP: 2280080989
Change-Id: Iaae56f748e6867f68ca5095b13dfbfed1412da55