[FIX] RenderManager: Keep track of focused element before preserving DOM
Original issue:
When a DOM element which contains the activeElement is moved to the
preserve area the focus is set to the body.
At the time the patcher wants to store the focus information,
the preserving has already happend and the currently activeElement is
not contained in the respective domRef anymore.
Thus the FocusHandler could not correctly apply the focus.

Snippix: #32532

PS01: solution idea
PS02: better implementation

BCP: 1970240933
Change-Id: I4c14d357a22021aea2e494bd14378bac4ddf4e6f