[FIX] sap.f.DynamicPage: Fix unnecessary scrollbar when footer is shown
Problem: When there is a table in the DynamicPage, which takes the full
height of the page and initially the footer is not displayed, if we
toggle the footer and show it, unnecessary scrollbar appears. This
happens due to a timing - the resize handler of the DynamicPage is
executed before the resize handler of the table, so that the DynamicPage
determines that a scrollbar should appear and the table does not have
the chance to adjust accurtately its height.

Solution: Does not prevent the rerendering of the DynamicPage, when the
visibility of the footer is changed. This way the timing issue is
prevented and the table occupies less space, allowing the footer to be
displayed, without showing the scrollbar.

BCP: 2270030291
Change-Id: I3d2d38182b5ab5de9983048b1b8ad4ddbdc6eb65