[INTERNAL][FIX] sap.m.MultiInput: State of the Tokenizer is now stored
We need to restore the state of the Tokenizer on MultiInput invalidation and
rerendering. That is why a new object property was added to the MultiInput's
instance (this._bTokenizerCollapsed) and private getter and setter were
created for better traceability and easier debugging.
This will allow for the MultiInput to restore correctly the state of the
Tokenizer when tokens are deleted and the width of the tokenizer is
recalculated.

CR-Id: 002075125800001472392022
BCP: 2270020567
Change-Id: Id88bce68a6610053adc3a7d75eaf1d57501a77c7