[INTERNAL] MDCTable: Remove column header menu from table control tree
If a control (or anything in the control) and it's (grand)parent are
invalidated at the same time, the child is not re-rendered directly, but
the core assumes it will be rendered as part of the (grand)parent
re-rendering.
Because the menu is a child of the table but is not rendered by the
table, the menu won't be re-rendered in such a case.

Change-Id: I2be9ec1c6003a77105438286c51e199ba3453b6b