[INTERNAL] mdc.Link: fixup after removal of XMLComposite
This will fix an issue which is caused by the observer of the Panel if
the _createContent is called before applySettings as the button
responsible for the p13n dialog is not created when we try to set it's
visibility.

Also added a more solid aggregation "forwarding" of the additional
content in the observer of the Panel.

BCP: 2270033296
Change-Id: I0b5d0fa95e80e921c913d6aaf3ced15f527701e2