[FIX] Demo kit: Doubling of version string in path is fixed
Calling ControlsInfo.loadData() was involving calls of lots
of unnecessary methods, some of which were mixing the idea of
ResourcesUtil. Ex: If there is an old version chosen and feature-flag
sap-ui-xx-unifiedResources were added, the source of resources inside
iframe was containing /1.**.**/ and calling
ResourceUtil.getResourceOriginPath was adding it once again to the
url the files were requested from.
With this change we only left what is essential for the URL prexies
of samples to be added to sap.ui.loader.config.

BCP: n/a

Change-Id: I1bbbbd26209b900eed424d7df1fa3df45de25025