[INTERNAL] Field: use unit-part type for ValueHelp
If the CompositeBinding for Currency/Unit has single types for the parts. Unse the type for the unit part for the value help. (As only the unit itseld is used in value help, nit the composite array.)
If no type is specified for the unit part, inside the default (String) will be used.

The type is needed to format the tokens in the value help dialog and in the DefineConditionPanel. (if shown)

Change-Id: Id46f30e71b101d333f1afa4fce4f65052a77caef