[FIX] sap.m.RangeSlider: Prevent tooltip reposition attemps if Slider is not rendered
The tooltips must be repositioned when the value is set programatically,
but there are cases where the Slider is not rendered yet and the repositioning
causes an error to be thrown - when the SliderTooltipContainer is looking
for its parent, which doesn't exist yet.

This is now prevented by checking if for the existence of the Slider.

Change-Id: Ied180872e055b13f257e42c29fa80d107230cfa1
BCP: 2270029757