[INTERNAL]Engine: cleanup state key externalization
The StateUtil used a hardcoded mapping to internalize/externalize
the state object to communicate with the Engine.

As the state representation is decided by the Controller, this logic
has been refactored and moved to the according controller instance.

The according controller can now overwrite the #getStateKey method
to define the key for the external representation.

Note: this cleanup also removes the cyclic depdency between Engine & StateUtil.

Jira: CPOUIFDENVER-1752
Change-Id: If066ed7bba848aa5760ad92b338624584105743e