[FIX] sap.uxap.ObjectPageLayout: Correct rounding issue
Problem:
Rounding issue with the height of the spacer on some resolutions
(e.g. after zoom-out to 90%):
When we change the spacer height to X
(e.g. by calling spacer.height(X + 'px'))
and check the resulting scrollHeight of the spacer,
the returned scrollHeight is 1px smaller than the
assigned height due to browser rounding.
This causes error in our logic that compares
the final height with the expected height

Solution:
Use jQuery.height to obtain the spacer height
(instead of using offsetHeight)
since jQuery.height also takes into account
the height assigned with css.
Also refactored the corresponding qunit
to remove the usage of fixed timeouts.

BCP: 2270033630
Change-Id: Ie10476147538d99362d5b9491209bef3e9e17b74