[FIX] sap.m.DynamicDateRange: Next/Last X days are no longer dependend on static options
Issue: Next 1 and Last 1 options were dependent on Yesterday and Tomorrow.
Solution: We have internal checks whether we include these options and if we
don't we don't parse to them. Same logic is applied with the formatter.
But because the formatter doesn't have access to the included options
we introduce a new bSkipCustomFormatting boolean variable.
If set to true, the strict formatting is skipped.

BCP: 2280024045

Change-Id: I04df09e0aa902fe93e175d7b0f447258ae139756