[INTERNAL] sap.uxap.ObjectPage: Add safe-check before calling doLazyLoading method
Problem: There are cases when the _adjustLayoutAndUxRules method is called before the oLazyLoading instance is initialized in onBeforeRendering hook. While it make sense to safe guard the whole function it may result in other layout issues.

Solution: Check if lazy loading is enabled and the lazy loading instance is defined before calling its APIs.

Change-Id: I41fe7f745c2f325d95d23e4985ec9186bf34312c
BCP: 2270043148