[INTERNAL] Integration cards: Header's accessibility is improved
The header now always has role="group".

An inner element inside the header is now receiving the focus, while the
focus outline is still drawn on the header element.

This helps the case when the header is interactive,
to avoid nesting a button element within a role="button" div.

Snippix: #664393

JIRA: BGSOFUIRODOPI-2743
Change-Id: I362f30fd6348f4aed5f9177b0b1d9e1299460994