[INTERNAL] core.Component: Correctly evaluate /sap.ui5/handleValidation
The "sap.ui5/handleValidation" property is now correctly evaluated for
instance-specific manifests.

Before this change the value of the "handleValidation" property was
evaluated from the standard Component manifest instead of the
instance-specific one.
The Component was then not correctly registered to the global
MessageManager. In this case no Messages were added, even if the
Component manifest set "handleValidation" to true.

Still, when calling getManifestEntry() on the Component instance,
the correct instance-specific manifest value is returned.

JIRA: CPOUI5FRAMEWORK-359
Change-Id: Ib6e7949367822258b6263b85753da4fcda6ef9b6