[FIX] sap.uxap.ObjectPageLayout: Prevent auto collapsing header
Problem: In some cases, when the user scrolls to the last section and
press the expand header button, the header gets expanded and
immediately after this, it gets collapsed.
This happens when the header has a fraction value height.

Solution: It seems that the spacer height is calculated wrong, with 1px
more than needed. When the header is expanded there are two scroll events:
first one is ignored, but the second one is processed and with just 1px
difference we calculate that the header should be snapped.

BCP: 2280068998
Change-Id: I8542f9f4f3dea732e300448b870d8e971121f48f