[FIX] P13nConditionPanel: Remove all should remove all conditions
Issue: When remove all is triggered all conditions are
removed from the Panel. If you close and open again
some conditions are available.

Solution: When removing the filters we're giving the index
of the filter which should be removed starting from
the 0 index. After the filter is removed
the indexes in the array are one less, so when we try to
remove a bigger index it doesn't exist anymore since the
array is having already less items

To be reviewed together with test in 217ade0e6c8768cb135040170a3a43612ec7fb17

Change-Id: I5255e5dad6b822fdf5faa14e3e4da1b63fbcba61
BCP: 2280087314
CR-Id: 002075125900000838122022