[FIX] sap.ui.model.odata.OperationMode.Auto: deprecation
With the "Auto" operation mode, the result of filtering and sorting may
differ depending on whether the data is filtered or sorted on the server
or on the client. This may lead to inconsistent behavior.

PS1: deprecated the operation mode and the threshold parameter in the
     ODataListBinding and in the ODataTreeBinding
PS2-6: Review comments

Change-Id: I6ed84850fdadee7e4665e04d2c0082e0f647c82d
JIRA: CPOUI5MODELS-759