[FIX] sap.m.Select: Fixed visual displacement of closing droppdown menu
Problem: When we change the ValueState of the sap.m.Select while we
close the drodpown (picker) menu, there is a visual displacement
of the latter.

Solution: We postpone the change of the ValueState and actually change
it after the drodpown menu is closed.

BCP: 002075129400001188692022
Change-Id: Ief3690752a67f02d7cdef4c2e4e62f349a9beef4
(cherry picked from commit 9b2bb2433dfab06c24253671e844ca5b9a6a15e3)
CR-Id: 002075125800002179752022