[INTERNAL] sap.ui.rta remove the Actions > SaveAs from the Manage App Variants
Due to the additional complexity of versions with RTA, selecting which version should be saved within the App Variant overview holds to much complexity.
The shortcut was removed an a switch to the actual app variant has to be done in order to select the save as in a WYSIWYG-style.

JIRA: CLOUDIOI-681
Change-Id: I3f31c6a6e611a6de7441e063616fb5a280b63c3b