[FIX] Demo Kit: Sample loading fixed
When opening a sample not following the Demokit
navigation, but opening a sample by adding its address
(entity/sap.m.DraftIndicator/sample/sap.m.sample.
SemanticPageDraftIndicator)
after the hash symbol in address bar, a race condition with still not
rendered DOM Ref of the iframe HTML object was hit. Now we removed
the optimisation logic that only changes the "src" property of already
rendered iframe, prior to creating the HTML control, every time a sample
is loaded, to make sure the sample is reloaded on every iteration.

BCP: n/a

Change-Id: I974f7ba221f45e89c4e19f3c47321d79e886a63c