[INTERNAL] sap.uxap.ObjectPageLayout: Merging title and actions
Problem: When there is only one ObjectPageSubSection in an
ObjectPageSection (SubSection is promoted), the title of the Section
inherits the title of the SubSection, and the title of the SubSection is
hidden. However, when there are actions (on SubSection level), they are
shown on the next line, which creates unnecessary white space.

Solution: Merging the title of the Section (inherited by the
SubSection), with the actions of the SubSection, throug CSS.

JIRA: BGSOFUIPIRIN-5042
Change-Id: Iaab9121285fd651dd0a0df2f3d2dd3fde4020054