[INTERNAL] MDC Chart: Reworked inner chart rendering
This commit introduces a new control: the ChartImplementationContainer.
Its main use-case is to provide a empty panel with a renderer to place the noDataText / sap.chart.Chart in.
This way, we can get rid of the VBox and all the limitations that come with it.
The MDC Chart itself has been changed to a CSS-Grid.

Still missing from this commit:
JS-Docu

Change-Id: I3ba3797883c6998c3196036c71ae9b16cc87dd88