[FIX] Demo Kit: Loading sample from previous version fixed
A loading error is appearing when "sap-ui-xx-unifiedResources" flag is
used and you try open a sample. Because of a recent development, we
pass all the query parameters from the Demo kit main window to the
iframe, so the flag parameter is also passed, also, when domain
of the iframe and main window matches the session storage is shared,
which causes some issues, when the resources address is constructed
and the version string was added two times. Now we only pass short-list
of parameters from the main window to sample iframe, to avoid unwanted
logic loaded.

BCP: n/a

Change-Id: Id19eb185e93b84d58f83926d1a1575969b8bbd7a