[FIX] sap.uxap.ObjectPageLayout: AnchorBar position is fixed
Problem: When the OPL is in iconTab mode and there are many Sections, if
the user selects one, which requires scroll to the right, and then
scroll down (such that the title is getting snapped), the AnchorBar
scroll is moved to the left and the selected section is not visible.

Solution: When the AnchorBar is moved in or out of the title area, its
scroll position is now restored, so the selected section is visible.

BCP: 2270040798
Change-Id: I8b07b8cd1a6661d2f1bf758b6cd277668949a889