[FIX] sap.m.Toolbar: Correct spacing for single item
Problem:
When the toolbar has a single child, the child is given an extra margin-right of 1.25rem.
That spacing comes from the base less and was added for Fiori1 and Fiori2 themes.
Fiori1 and Fiori2 specify margins before the first and after the last child items.

In Fiori3 and Horizon, however, the space after the last child comes from the padding
of the Toolbar itself => the margin-right of the last child is redundant.

Solution:
Overwrote the base style in the context of Fiori3 and Horizon.

BCP: 2270065006
Change-Id: I7c1196fda1ecfde772460ae7d28686fcf3e59881