[INTERNAL][FIX] ColumnMenu: Enable visibility option for QuickActions
Specifying a QuickAction's "visible" property is now taking effect and setting said property to "false" will hide the QuickAction from the ColumnMenu.

If there are no active/visible QuickActions, the QuickAction container is not going to be rendered.

Moved Entry#setVisible to ItemBase#setVisible, because previously Entry#setVisible contained Item-specific logic, which interferred with setting a QuickAction's visibility.

Adding/Removing items and quick actions after the menu was opened, will now also have an effect in the ColumnMenu.

Additionally, added Unit tests for setting visibility in the ColumnMenu.

Change-Id: I5aad04c5dfb8d276ac66e6de671e5d65a5346071