[INTERNAL] sap.ui.rta remembers the translation state in case a translatable change was saved
To ensure a proper enablement of the translation button, the state of the button must kept after a in-session save happened.
This commit already ensures a save done on the "download translation" option keeps the translation button active, activating a version will be added in a separate commit.

JIRA: CLOUDIOI-242
Change-Id: I9724eff830ff170321e297e1c545c607dcfbf780