[FIX] sap.m.Toolbar: Correct visualization in DynamicPage
Problem:
With horizon theme, toolbars nested in the content aggregation
of the DynamicPageTitle have an unwanted bottom border.

Previous solution:
The unwanted bottom border was previously removed with CSS in the
context of DynamicPageTitle. However, that CSS was not applied
in horizon, because horizon sets greater weight to the original
selector => the overriding selector did not apply.

Solution:
Increased the weight to the overriding selector (used in the
previous solution) to match the increased weight of the original
selector in horizon
(The increase in the weight of the original selector in the context
of horizon seems unproblematic and reasonable and is left as it)

BCP: 2270079633
Change-Id: I8f440eb52fffd1977e41c8a60ff84915519d8aa6