[INTERNAL][FIX] sap.uxap.ObjectPageLayout: subSectionEnteredViewPort
Problem: When _preloadSectionsOnBeforeScroll is called, the
subSectionEnteredViewPort event is fired before the unstashing of the
SubSections, which breaks FE functionality for enabling stashing of
SubSections. At the moment the event is fired, the SubSections are still
stashed and there content is not added as an aggregation.

Solution: The problem comes from a delay, implemented for a destkop
mode, which is meant to not block the animation scrolling. Now we are
firing the subSectionEnteredViewPort event just after the SubSections
are unstashed (with a delay).

BCP: N/A
Change-Id: Id37459536cef3ad3544550ad4683c161dea2a96c