[INTERNAL]SelectionPanel: improve screen reader output for 'active' state
By default the table will only read out the column header texts
when accessing cell items.

The 'SelectionPanel' allows to use a second 'active' column which can be
used to indicate whether a value has been filled.

As the visual impact is not given in the screen reader output,
an additional InvisibleText has been added to provide further
information in the screen reader usage.

Jira: CPOUIFDENVER-1745
Change-Id: I59ae0025fe6a7e7cf01b150f8aaae427c138a78c