[INTERNAL][FIX] sap.ui.rta: Avoid checking technical properties during reveal evaluation
When a control has bindings against technical properties like RAP
feature control properties, those are not always filtered by the
Delegate.
Depending on the order in the Delegate, these technical properties might
be evaluated instead of the actual content property.
This can be avoided by first sorting by the hideFromReveal attribute,
which doesn't guarantee that the content property is selected but at
least prioritizes it compared to technical properties which always have
the flag set to true.

This is a temporary solution which might fail if a control has content
bindings against multiple non-technical properties of which one should
not be revealed.

Change-Id: Icbb1acba42b5466cf85a08cfa5c868f04265e658
BCP: 2270044840
BCP: 2270080416