[INTERNAL] m/Table: Column and ColumnResizer issues are fixed
- Column should cleanup the initialOrder attribute when it is removed
from the table since this attribute is only a cache of the column index
for a better rendering performance.
- In case of a column is removed and inserted to the table again the
position might change. The autoPopin feature runs in onBeforeRendering
before the initialOrder is correctly aligned in the renderer.
- ColumnResizer handle should get hidden when it is hovered out
- Headers association seems enough and additional ariaLabelledBy
association from the column header to the cell is not removed.
- Unit tests are adapted accordingly.

BCP: 2270071078
BCP: 2270066965
BCP: 2170316617

Change-Id: I849b482202d6726535d914c96f108b4faa14d8fb