[FIX] sap.m.MultiInput: Single cell is now pasted as text
For Windows an extra string was attached at the end of the clipboard
data string, which is used for the paste functionality.
This is the reason for converting a single cell into a token, which is wrong
according to our documentation.
Also '\t' is also added to the regex used for converting the clipboard data into an array,
since it is also a separator symbol.

Change-Id: I81d5f2667c3efb5c463f6f97facf237a291e6e00
BCP: 002028376700032754172022
CR-Id: 002075125800002637752022