[INTERNAL] sap.m.Select: Fixed incorrect hover style in table cell
Problem: The Select control has incorrect style on hover when it's in
table cell on horizon themes.

Root Cause: There are redundant old styles which are outdated in the
Select's base less file.

Solution: Delete the redundant styles in order to mimic the looks of
other similar controls (like sap.m.ComboBox).

BCP: 2270086930
Change-Id: I0be0e15a2a8d40d0ed98882b23133fbc94fdf5f9