[FIX] sap.uxap.AnchorBar: Switch to HierarchicalSelect optimized
This reverts commit 732dface6efefc78c8514d6745a006d6990d4c9f.

Reason for revert: After I889e337ddc5d3df579c2531285a15c163cd53ea7 this change is unnecessary. The classes sapMBarFirstVisibleChild & sapMBarLastVisibleChild are no longer presented and the original issue is no longer reproducible. Also, it brings new complexity as after the invalidation of the anchor bar the computation of the sections info (_computeBarSectionsInfo) is not called at the right time (right after the sapUxAPAnchorBarScrollContainer is shown).

Change-Id: Ie0799eac4aeda71270b7b4edfcf7a7b1c7dc252d
BCP: 002075129400000511102022