[FIX] sap.ui.unified.FileUploader: busy state and hover interaction properly applied
The sap.ui.unified.FileUploader is properly displayed,
when its instance is added as contents to another control,
which is set to busy 'true' state.

Issue: when the Browse button was hovered, it would not get all styles applied
as the element over it was hit first.
Solution: The button element is using a higher z-index in order to be hit
by the hovering motion.

BCP: 2170304073
BCP: 2180210567
BCP: 002075125800002681512022
Change-Id: I4736a798ada0740dde1ded3032d5ce105a566db2
CR-Id: 002075125800002681512022