[INTERNAL] DataStateIndicator: Do not expose logical owner but parent
This change is in relation to the change below
I7df253a13587e6e3f0699e749595e3932206e51f
DataStateIndicator should return the parent control instead of the logical owner.
In the change mentioned above the fix was done when processing the data states, but not when the user clicks on the "Filter items" link in the message strip.
This has now been adapted to the return the parent control as the parameter to the filter function.
BCP: 2270089268

Change-Id: I465f5dd4d55e1cc80c8daaaf308de3ae3c02f766
(cherry picked from commit 80a42eed43c4fd47709807b5f04a0dd030c21c41)
CR-Id: 002075125800002805742022