[INTERNAL] AddTableColumn: fix ResponsiveTable changeHandler
The 'newCellSelector' in the RevertData object was always set as isIdLocal as 'true'. The isIdLocal configuration prefixes the component id to the control id which is not always correct. The issue reported in the incident below, the 'undo' action was not worked as expected, since the 'revertAdditionalControls' function was not able to retrieve the expected cell template and due to this the cell could not be removed from the template.
To fix the issue we now make use of the helper function sap.ui.core.util.reflection.BaseTreeModifier.getSelector, which helps in determining whether isIdLocal should be set to true or false.
Unit tests are adapted accordingly.
BCP: 2270083878

Change-Id: Iebe128c3f86f26d2c8218135bdbaafe8e3a2fb1b
CR-Id: 002075125800002851582022