[FIX] sap.f.TargetHandler: Close all Dialogs when navigating back
Problem: When FlexibleColumnLayout is used and the user navigates back
with the back browser button, Dialogs remain opened.

Solution: The problem comes from the FCL functionality, which considers
a navigation only when a new page (not among the current begin, mid or
end column's current pages) is opened.
Now when navigating back, all Popovers and Dialogs are closed.
However, still keeping the functionality for not closing when navigating
forward, as there are some cases, when navigating from list items,
placed in a Popover/Dialog, should not close it.

BCP: 2280065097
Change-Id: I1eb052bd59f62abccd78d0cee9424153aefa95a5