[INTERNAL] Field: Timezone brings error
The DateTimePicker requires the timezone as IANA timezone ID. So it must not be formatted by the DateTimeWithTimezone type to the locale language.
Therefore a new target type for the ConditionType is introduced to allow transfering the raw values in formatting and paring. So, if sap.ui.mec.raw is set in Binding as targetType, the ConditionType is working with raw values.
For CompositeTypes sap.ui.mec.raw:0, sap.ui.mec.raw:1... can be used to define the part of the needed raw value.

Change-Id: I5fe92492bb10de9c6d637ff2d64c243e2154209a