[INTERNAL] JSDoc: add null/undefined to @returns when documented in text
In some cases, a return type of null or undefined was only mentioned in
the method or return value description, but not in the type expression.
This resulted in incomplete TypeScript signatures.

In a few other cases, the documented return type did not match the
implementation.

This is a follow-up to https://github.com/SAP/openui5/issues/3515.

Change-Id: I205a7bc9e8e34c8574d39889114c461d16017b69