[INTERNAL] sap.m.Slider: Fix unstable qUnit test
There was a test that was fired 100 times in a loop in order to test the tapping event multiple times.
After changes in the Slider's code from change e501c7fb094051a208698744d5404d1a6204e48f
the Slider is now getting re-rendered on tap. This causes the test to rerender the Slider many times in a loop
which leads to unexpected results.

The loop is now removed and the test is being run only one time.

Change-Id: I276f1257041ae5ddc3ca1714a9ca97f684ae0888