[INTERNAL] sap.m.DynamicDateRange: visual tests stabilised
Issue: Some of the tests were skipped.

Reason: The JSDate object was mocked in the spec file to return
solid date in order to build stable visual tests. This is needed,
because the DynamicDateRange picks relative to the current date.
Because of this mock, the Calendar inside the DDR was not
initialised properly, because the mocking had been done too late.

Solution: now the mocking is done before building the test html page.
Change-Id: I842c4047ba44498b93491253d3babb2431ad6e18