[FIX] GrowingEnablement: asynchronous items rendering issue is fixed
 - The rendering of items, applying the growing chunk, has been made
asynchronous in the previous version to support the merging cells in
ODataV4 since property binding resolves async and the cell merging
happens during the rendering time.
 - This changes now limits the async rendering behavior to the ODataV4
and cell merging.
 - In addition if the sync item deletion or insertion is inevitable then
chunk handling gets ignored and we invalidate the control to apply all
changes in an async rendering cycle. This also ensures rendering the
items in the correct order.
 - As a performance optimization, if a control has items but no DOM
output then we do not need to destroy all items since we can reuse them
at the next binding update.
 - Unit tests are adapted accordingly.

BCP: 2280083462
BCP: 2270108409
BCP: 002075129400003845712022

Change-Id: I819da7d27a50cf64e8738b14bd83c75f8d7c9c03