[FIX] sap.f.DynamicPage: Fix expand/collapse header button
Problem: When header of DynamicPage is destroyed and new one is set, the
expand/collapse header buttons do not work.

Solution: Destroying the header now detaches all event listeners and
resets all properties of DynamicPage, related to the current state of
the header, so that if new one is set, it will receive all event
listeners.

BCP: 2280083433
Change-Id: Iefa9b3d5cd05a11de00a182a93e555a2dd03a326