[FIX] sap.m.ProgressIndicator: Fix cutting of items in a table
Problem: With sap_horizon theme the ProgressIndicator has height, which
stretches table's cells. As the table has fixed height of its cells,
bigger ProgressIndicator breaks the calculation logic of how many rows
can be shown and some of them are cut off.

Solution: There are base styles of the ProgressIndicator, when it is
placed in a table. However, in sap_horizon theme they do not work,
because we use paddings, instead of margins. Now these styles are
implemented for sap_horizon theme too.

BCP: 2280084642
Change-Id: Ie07152419f08fb657c6615fc679cb19f3861e715