[INTERNAL] Table: fix accessibility issue for focusable control in popin area
The column headers are cloned which get a new id, when rendered in the popin area. Due to this controls rendered within a layout control do not have a valid ariaLabelledBy association as the original header id is not available in the DOM.
To solve this issue, the table now renders a hidden DIV element which contains the column headers that are moved to the popin area. The original column header is rendered in this DIV element.
This ensure valid ariaLabelledBy associations and screen readers can announce the expected information.
BCP: 2170290335

Change-Id: Iecc04f358230c30ef90a40a89e5dd8bd6cf100a6