[FIX] sap.m.MessagePopover: Focus handling corrected
_restoreFocus method of the MessageView is called after each navigation
of the NavContainer. This ensures that the focus is correctly applied
each time.
When the MessageView is used in a MessagePopover the focus handling is not correctly applied,
as the afterNavigate is not called at first rendering of the popover.
With this fix we ensure that the focus is applied in the MP scenario.

Change-Id: If5acd6015ae65a6807de1301b91d96895f1a3a38
BCP: 2270031987