[FIX] sap.m.RadioButton: Selection in table with many radio buttons is now correct
Problem:
The setter for property "selected" relies on another property - "groupName" to deselect all other buttons in the same group.
But the "groupName" property might not be updated yet, which results in updating based on the old "groupName".

Solution:
Ensure that "selected" and "groupName" are both set, by using the onBeforeRendering hook.

Additionally, the sap.m.RadioButtonGroup is now updated to handle the new behavior of the RadioButton.

Snippix: 427549, 27978, 97394, 32864, 700623
BCP: 002075129500003326492022
Change-Id: I94103f7ac94599aa41e45730ab7177dddfcb76d8