[FIX] sap.m.DynamicDateRange: DateTimeRangeSelection now works with timezone
As we use the DateTimePicker as standalone control in this specific option,
and we use the getDateValue method. This method returns the date in the
timezone configured, not in local timezone. That's why we set as configuration
the local timezone, so the control always return the date in local timezone.
The issue was that the timezone shifting was done twice, once in DTP and once in DDR.
Now it will only be done in DynamicDateRange.

BCP: 2280090355

Change-Id: I964f42e54070f81c5828e74df03e7ac3b394d007