[INTERNAL][FIX]sap.ui.dt,sap.ui.rta: OP rename scroll issue
In some cases, during the rename operation a scrolling can happen
basically at the same time, causing the editable field to be placed
on the wrong position. E.g. focus on OP Section after context menu is
closed.

This fix ensures that, if an overlay has scrolling, whenever the
scrolling gets synced the position of the editable field is calculated.

Change-Id: Id12856c4cae9db59b0d45a051c4a1c112ae2ca51
BCP: 2270083916