[INTERNAL] mdc.Table: adding column after the table (responsive table) is grouped should not throw an error
GroupHeaderListItem controls are inserted in the items aggregation of the responsive table when groupped. This control does not have the "cells" aggregation, which lead to an error while adapting the cell template after a column is added via personalization. This error is now fix by ignoring the GroupHeaderListItem control when adapting the cell template.
BCP: 2280085617

Change-Id: Iea2edd990a6536056803197911b79843f537fd14
CR-Id: 002075125900001115972022