[INTERNAL] mdc.Table: Create Opa Page Object and Opa tests
This change includes by side of the Opa Page Object and Opa Tests for the MDCTable
also some refactoring of the existing Actions and Assertions done for the MDCTable.
Some of them are not relevant for the public section and some relies on properties of the SmartTable
or even not presented in the MDCTable at all.

JIRA: CPOUIFTEAMB-1845
Change-Id: Idc9d870c01d33dd9f3aa9a5b7ef65c64cea434e4