[INTERNAL] DemoKit: 404 page now redirects to known resources
Problem:
There are links to Demo Kit which are actually leading to content but,
because they are part of removed versions, these links lead to 404 page.

Solition:
Enhance logic in the 404 page to understand when URLs are leading
to content in Demo Kit and not leading to a code file like a library.js
for example. If such URL is detected the 404 page can redirect to the
correct page in the Demo Kit.

JIRA: BGSOFUIPIRIN-5748
Change-Id: I57cb31185615d465b23b296a23d744d5b7ff75bb