[FIX] sap.uxap.ObjectPageLayout: Allow RTA revert moved headerContent
Problem:
The "revert" action does not work for items from the "headerContent"
aggregation of ObjectPageLayout that have been  moved using RTA

The problem occurs because the "revert" action depends on correct
parent-child relationship.
But for any item of the "headerContent" aggregation of
ObjectPageLayout, their getParent() function does not return the
ObjectPageLayout instance, but the instance of the internally
created ObjectPageHeaderContent container object.

Solution:
Ensured that getParent() function returns the
ObjectPageLayout instance for all items of the "headerContent"
aggregation of ObjectPageLayout

Change-Id: Ia4bf932d8d400fb07e0dd86e06934e4b828f93eb
Snippix: 169560
BCP: 2180311371