[FIX] sap.m.Toolbar: Padding adjusted for Horizon theme
Problem: The left padding of the Transparent Toolbar is less than the VD spec. As a result, the title of the Toolbar appears misaligned with the first column of sap.m.Table.

Solution: Increase the padding of the Toolbar container with 0.25rem to align to the specification of 1rem.

BCP: 2280101796
Change-Id: Id7c6b534603333c31e335c6ed5bfd0a427142017